+++ lib/sv.c
@@ -93,7 +93,7 @@ int sv_measure_l(comedi_sv_t *it,double *data)
 	val=malloc(sizeof(*val)*it->n);
 
 	chan=CR_PACK(it->chan,it->range,it->aref);
-	
+
 	t.subdev=it->subdevice;
 	t.mode=0;
 	t.flags=TRIG_DITHER;
@@ -108,7 +108,7 @@ int sv_measure_l(comedi_sv_t *it,double *data)
 	for(n=0;n<it->n;){
 		t.data=(void *)(val+n);
 		t.n=it->n-n;
-		i = comedi_ioctl(it->dev->fd, COMEDI_TRIG, (unsigned long)&t);
+		i = comedi_ioctl(it->dev->fd, COMEDI_TRIG, &t);
 		if(i<=0){
 			ret=i;
 			goto out;
@@ -144,7 +144,7 @@ int sv_measure_s(comedi_sv_t *it,double *data)
 	val=malloc(sizeof(*val)*it->n);
 
 	chan=CR_PACK(it->chan,it->range,it->aref);
-	
+
 	t.subdev=it->subdevice;
 	t.mode=0;
 	t.flags=TRIG_DITHER;
@@ -161,7 +161,7 @@ int sv_measure_s(comedi_sv_t *it,double *data)
 	for(n=0;n<it->n;){
 		t.data=val+n;
 		t.n=it->n-n;
-		i = comedi_ioctl(it->dev->fd, COMEDI_TRIG, (unsigned long)&t);
+		i = comedi_ioctl(it->dev->fd, COMEDI_TRIG, &t);
 		if(i<=0){
 			ret=i;
 			goto out;
